[Refactor] refactor element name and separate file
요소 이름을 변경하고 파일을 분리하였다.
