+++ pokemon/index.html
@@ -18,6 +18,6 @@
 
     <section class="imgContainer"></section>
 
-    <script src="index.js"></script>
+    <script type="module" src="app.mjs"></script>
   </body>
 </html>
